# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "BudgetBuddy API Doc",
        "version": "1.0.0"
    },
    "paths": {
        "/api": {},
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login user",
                "description": "Login user",
                "operationId": "84e0696e7526b45c02213e3b6d12d290",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "manoj@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "expires_at": {
                                            "type": "integer",
                                            "example": 1717932517
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid Credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-detail": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "User Detail",
                "description": "Authenticated User's Detail",
                "operationId": "a913ea8289da989e4beb95a907a16f60",
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "success",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "User's ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "User's name",
                                                    "type": "string",
                                                    "example": "Test"
                                                },
                                                "email": {
                                                    "description": "User's email address",
                                                    "type": "string",
                                                    "example": "test@gmail.com"
                                                },
                                                "email_verified_at": {
                                                    "description": "Timestamp of when the user's email was verified",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-09T09:00:24.000000Z"
                                                },
                                                "created_at": {
                                                    "description": "Timestamp of when the user was created",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-09T09:00:24.000000Z"
                                                },
                                                "updated_at": {
                                                    "description": "Timestamp of when the user was last updated",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-06-09T09:00:24.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "description": "Error message",
                                            "type": "string",
                                            "example": "Invalid Credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Logout user",
                "description": "Logout user",
                "operationId": "97dbfb020c5e67c3060a3ef69cae867f",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid Credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "edcb2984d0169157be9a16c9e1a326d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Manoj"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "manoj@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "201"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "500"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Login": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "User's email",
                        "type": "string"
                    },
                    "password": {
                        "description": "User's password",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "description": "User's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User's email",
                        "type": "string"
                    },
                    "password": {
                        "description": "User's password",
                        "type": "string",
                        "format": "password"
                    },
                    "password_confirmation": {
                        "description": "User's password confirmation",
                        "type": "string",
                        "format": "password"
                    }
                },
                "type": "object"
            }
        }
    }
}